% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage}
\format{
## `who`
A data frame with 15 rows and 2 variables:
\describe{
\item{DISTANCE}{distance from fire station in miles}
\item{DAMAGE}{x 1000 dollars in damage}
\item{year}{Year}
...
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire
}
\description{
A data set collected concerning home damage from fire
Report ...
}
\keyword{datasets}
