% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the sample}

\item{fun}{function being called}

\item{alpha}{the confidence interval}

\item{cx}{cx}

\item{...}{...}
}
\value{
A bargraph plot containing the fun value mean and lower/upper as well as a large table of data under different sections
}
\description{
Title
}
