% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of sets of iterations}

\item{iter}{number of iterations}

\item{lambda}{lambda value}

\item{...}{...}
}
\value{
a histogram plot of the sample mean of the created poisson samples
}
\description{
Title
}
