% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhypersim.R
\name{myhyper}
\alias{myhyper}
\title{Hyper Geometric Simulator}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{The number of times the simulation iterates}

\item{N}{The total number of items}

\item{r}{The amount of whites(1's)}

\item{n}{The number of times a "marble is drawn"}
}
\value{
A barplot showcasing the probability of different number of successes based off iter number of iterations where there are N total marbles, r white marbles, and n marbles drawn
}
\description{
Hyper Geometric Simulator
}
\examples{
\dontrun{myhyper(iter=1000,n=19, N=20,r=12)}
}
