% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynarea.R
\name{mynarea}
\alias{mynarea}
\title{@title mynarea}
\usage{
mynarea(a, mu, sigma)
}
\arguments{
\item{a}{the right bound}

\item{mu}{mu(mean)}

\item{sigma}{sigma(sd)}
}
\value{
A Plot of a curve and a shaded section between -inf to a with the calculated area between returned
}
\description{
@title mynarea
}
\examples{
\dontrun{mynarea(4, 10, 3)}
}
