% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{@title myncurve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{the right bound}

\item{mu}{mu(mean)}

\item{sigma}{sigma(sd)}
}
\value{
A Plot of a curve and a shaded section between -inf to a with the calculated area between returned
}
\description{
@title myncurve
}
\examples{
\dontrun{myncurve(4, 10, 3)}
}
