% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Title}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{starting x value}

\item{delta}{delta rate}

\item{llik}{function}

\item{xrange}{range of x values}

\item{parameter}{lambda, p, etc..}
}
\value{
a graph and list of x,y values
}
\description{
Title
}
