% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets Calculator}
\usage{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of seats in plane}

\item{gamma}{probability of overbooking}

\item{p}{probability of show}
}
\value{
The number of seats a plane should sell without there being too many passengers to seat
}
\description{
Number of Tickets Calculator
}
